+++ plone/app/contenttypes/behaviors/collection.py
@@ -31,6 +31,7 @@ class MetaDataFieldsVocabulary(object):
         ]
         return SimpleVocabulary(items)
 
+
 MetaDataFieldsVocabularyFactory = MetaDataFieldsVocabulary()