+++ plone/app/contenttypes/tests/test_image.py
@@ -13,8 +13,9 @@ from zope.component import createObject
 from zope.component import queryUtility
 from zope.interface import alsoProvides
 
+import io
 import os.path
-import unittest2 as unittest
+import unittest
 
 
 def dummy_image(filename=u'image.jpg'):
@@ -141,7 +142,7 @@ class ImageFunctionalTest(unittest.TestCase):
         self.browser.getControl(name=widget).value = 'my-special-image.jpg'
         image_path = os.path.join(os.path.dirname(__file__), 'image.jpg')
         image_ctl = self.browser.getControl(name='form.widgets.image')
-        image_ctl.add_file(open(image_path), 'image/png', 'image.jpg')
+        image_ctl.add_file(io.FileIO(image_path), 'image/png', 'image.jpg')
         self.browser.getControl('Save').click()
         self.assertTrue(self.browser.url.endswith('image.jpg/view'))
         self.assertIn('My image', self.browser.contents)
@@ -157,7 +158,7 @@ class ImageFunctionalTest(unittest.TestCase):
         self.browser.getControl(name=widget).value = 'my-special-image.jpg'
         image_path = os.path.join(os.path.dirname(__file__), 'image.jpg')
         image_ctl = self.browser.getControl(name='form.widgets.image')
-        image_ctl.add_file(open(image_path), 'image/png', 'image.jpg')
+        image_ctl.add_file(io.FileIO(image_path), 'image/png', 'image.jpg')
         self.browser.getControl('Save').click()
         self.assertTrue(self.browser.url.endswith('my-special-image.jpg/view'))
 
@@ -173,9 +174,9 @@ class ImageFunctionalTest(unittest.TestCase):
         self.browser.getControl(name=widget).value = 'This is my image.'
         image_path = os.path.join(os.path.dirname(__file__), 'image.jpg')
         image_ctl = self.browser.getControl(name='form.widgets.image')
-        image_ctl.add_file(open(image_path), 'image/png', 'image.jpg')
+        image_ctl.add_file(io.FileIO(image_path), 'image/png', 'image.jpg')
         self.browser.getControl('Save').click()
-        self.browser.getLink('Click to view full-size image…').click()
+        self.browser.getLink('Click to view full-size image').click()
         self.assertTrue(
             self.browser.url.endswith('image.jpg/image_view_fullscreen')
         )